requirements.txt
asgiref==3.5.0
Django==4.0.4
django-appconf==1.0.5
django-imagekit==4.1.0
djangorestframework==3.13.1
pilkit==2.0
Pillow==9.1.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
